requirements.txt
geopy==2.4.0
numpy==1.24.3
pandas==2.0.2
python_dateutil==2.8.2
Requests==2.31.0
